scripts/requirements.txt
# Core - required for Bronze tier
watchdog>=4.0.0

# Gmail watcher (Silver tier)
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

# WhatsApp watcher (Silver tier)
playwright>=1.40.0

# Utilities
python-dotenv>=1.0.0
